+++ src/engine/core_systems/SysDrawAnimatedSprite.ts
@@ -1,10 +1,10 @@
-import CompTransform from '../engine/core_components/CompTransform';
-import EcsManager from '../engine/core/EcsManager';
-import System from '../engine/core/System';
-import CompDrawable from '../engine/core_components/CompDrawable';
-import PhaserContext from '../engine/core/PhaserContext';
-import Archetype from '../engine/core/Archetype';
-import CompAnimatedSprite from '../engine/core_components/CompAnimatedSprite';
+import CompTransform from '../core_components/CompTransform';
+import EcsManager from '../core/EcsManager';
+import System from '../core/System';
+import CompDrawable from '../core_components/CompDrawable';
+import PhaserContext from '../core/PhaserContext';
+import Archetype from '../core/Archetype';
+import CompAnimatedSprite from '../core_components/CompAnimatedSprite';
 
 export default class SysDrawAnimatedSprite extends System {
   private _archetype = new Archetype(
